Add proper oglContext for shared management between context. They should not be thread-local but context-sensative. LRUCache is also fixed. Add some /MP conpile options

+++ OpenGLUtil/BindingManager.cpp
@@ -12,12 +12,12 @@ GLuint TextureManager::getID(const oglTexture& obj) const
 	return obj->textureID;
 }
 
-void TextureManager::innerBind(const oglTexture& obj, const uint8_t pos) const
+void TextureManager::innerBind(const oglTexture& obj, const uint16_t pos) const
 {
 	obj->bind(pos);
 }
 
-void TextureManager::outterBind(const GLuint pid, const GLuint pos, const uint8_t val) const
+void TextureManager::outterBind(const GLuint pid, const GLuint pos, const uint16_t val) const
 {
 	glProgramUniform1i(pid, pos, val);
 }
@@ -28,12 +28,12 @@ GLuint UBOManager::getID(const oglUBO& obj) const
 	return obj->bufferID;
 }
 
-void UBOManager::innerBind(const oglUBO& obj, const uint8_t pos) const
+void UBOManager::innerBind(const oglUBO& obj, const uint16_t pos) const
 {
 	obj->bind(pos);
 }
 
-void UBOManager::outterBind(const GLuint pid, const GLuint pos, const uint8_t val) const
+void UBOManager::outterBind(const GLuint pid, const GLuint pos, const uint16_t val) const
 {
 	glUniformBlockBinding(pid, pos, val);
 }
